requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6

SQLAlchemy==2.0.36
pymysql==1.1.1
alembic==1.14.1

httpx==0.27.2

python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.7.1

python-jose==3.3.0
passlib[bcrypt]==1.7.4

celery==5.4.0
redis==5.0.8
